requirements.txt
flask
diffusers
transformers
accelerate
safetensors
huggingface_hub
Pillow
# Install cog-sdxl utilities directly from GitHub:
git+https://github.com/replicate/cog-sdxl

